requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.10.4
argparse==1.2.1
beautifulsoup4==4.3.2
boto==2.38.0
cov-core==1.15.0
coverage==3.7.1
decorator==3.4.2
html2text==2015.6.21
itsdangerous==0.24
ner==0.1
networkx==1.9.1
py==1.4.30
pylru==1.0.9
pytest==2.7.2
pytest-cov==1.8.1
python-coveralls==2.5.0
python-Levenshtein==0.12.0
# stanford-corenlp-pywrapper==0.1.0 not in pip. 
urllib3==1.10.4
wsgiref==0.1.2
